signin: removes dynamic client creation.
This is not completely possible as there doesn't seem
to be a way of making the user's own CLI client part of the audience
of the tokens generated for Hooklift's official CLI app, so the user
app can refresh tokens. If security concerns come up later on, we can
revisit this.