Finally implements a secured SSO authorization.
We implemented the concepts of audiences in our authorization
server so that we could get SSO tokens from native clients.

It turns out, audiences support when requesting grants started to be specified at
https://tools.ietf.org/html/draft-tschofenig-oauth-audience-00
but it expired and didn't seem to make it through the standarization process. Despite this,
some well known identity providers such as Google Accounts implements the concept of
cross-client identity, which solves the same problem.

That above said, this change creates a CLI native OAuth2 client per user.
There is no client credential leaking among users and we can suspend rogue
clients without affecting all the users. Additionally, we can use the ID and
Access tokens to access other Hooklift services without asking the user to re-authorize
the client.